0.7.1.3.9

+++ AMI/Neitsillia/Areas/Arenas/ArenaModifier.cs
@@ -52,7 +52,7 @@ namespace AMI.Neitsillia.Areas.Arenas
             }
         }
 
-        public bool WaveProgress(int score)
+        public bool WaveProgress(long score)
         {
             CurrentScore += score;
             wave++;
